WIP gtg

+++ src/bData/bRect.cpp
@@ -8,10 +8,10 @@
 SDL_Rect bRect::sdl_Rect() {
 
 	SDL_Rect rect;
-	rect.x = (int)x;
-	rect.y = (int)y;
-	rect.w = (int)width;
-	rect.h = (int)height;
+	rect.x = x;
+	rect.y = y;
+	rect.w = width;
+	rect.h = height;
 
 	return rect;
