+++ src/bData/bRect.h
@@ -7,13 +7,14 @@
 #define bRect_h
 
 #include <SDL2/SDL.h>
+#include <cstdint>
 
 struct bRect {
 
-	double x;
-	double y;
-	double width;
-	double height;
+	uint32_t x;
+	uint32_t y;
+	uint32_t width;
+	uint32_t height;
 
     SDL_Rect sdl_Rect();
